Handle .cabal conditionals
I was developing a library that needed bytestring-builder---which is an
additional package for older versions of bytestring (like the one in ghc
7.6), but whose functionality is included with the one included in ghc
>= 7.8.  So I had a conditional in my .cabal file.

The recent changes to get-cabal-configuration.hs caused flycheck to
start throwing warnings like:

    Suspicious state from syntax checker haskell-ghc: Checker haskell-ghc returned non-zero exit code 1, but no errors from output: <command line>: cannot satisfy -package bytestring-builder
        (use -v for more information)

    Checker definition probably flawed.

I tracked this down to get-cabal-configuration not handling conditionals
properly, and thus trying to reference a package that wasn't going to be
installed.

This change borrows code from Cabal in order to properly handle
conditionals---most of which really isn't about handling conditionals,
it's about making sure that the test and benchmark suites are enabled so
their dependencies will be picked up as well.  And then we also have to
be sure to filter out any references to our own package.