Add a small helper for producing CPP switches
Cabal has no commitment to a public api, and backward compatibility for
some changes are not representable purely at the Haskell level.  The
only possible route for supporting older or newer versions requires some
CPP-level hackery.  That means we need something that can feed us the
flags we need to feed to our helper.  This is intended to be that tool.