Conditionalize code for different cabal versions
The big break here is in cabal-1.22, where a parameter to
finalizePackageDescription were changed to take a different (new) type.

Because it wasn't present in older versions, we couldn't even resort to
typeclass abuse or some such as a workaround, we had to go with CPP (and
needed an outboard helper at that.