Ignore sandbox configuration when not present
I don't use cabal sandboxes for haskell development, so when I updated I
started getting the error:

  Error while checking syntax automatically: (error "Value (nil) of flycheck-ghc-package-databases for option \"-package-db\" is not a list of strings") [10 times]

This was because the prior code would create a single-entry list
of (nil).  This change simply skips monkeying with the sandbox config
stuff entirely if there isn't one, which seems like the correct course
of action.