Support envrc
`envrc` is an emacs package that hooks into `direnv` by changing
certain variables, namely `process-environment` and `exec-path`, to be
buffer-local, and updating them with variables from `direnv`. Among
other things, this can be useful to make build tools available within
certain directories.

`with-temp-buffer` switches to a different buffer, in which these
buffer-local variables are no longer set, instead "reverting" to their
global defaults. This defeats the mechanism employed by `envrc`.

Explicitly copy these variables in when we are using
`with-temp-buffer` to run a command, to better support envrc.